Subscribe frontend to BookingResponse
Removed extraneous lines of code from the publish method, converted the code in mounted into a subscribe method, set subscription to BookingResponse and publish to BookingRequest. Only one instance of the client is now being created. Set up filter to detect if the BookingResponse is linked to the BookingRequest just submitted.

Resolves #7.