Display dentist office name on booking page
Dominique and Victoria have added functionality in Booking so that it now subscribes to Dentists and free-slots, qos was also updated. A populateDentistArray method was created and now the names of all dentists offices are displayed on booking page. This will be refined to only display the information concerning the specific office.

Resolves: #21