Integrate error handling and edit booking form
Dominique, Negin, Shonaigh, & Victoria integrated the error handling for the booking request form and changes to how the request information is being passed into the publish method into the branch for handling simultaneous bookings. We also made slight changes to the logic for comparing the userId and requestId in the payload string and the booking request for checking for double bookings.

Resolves: #22, #23, #24