Edit styling of home page
Negin & Shonaigh removed the About Us page, added a logo to the nav bar, increased the size of the site name. We also made changes to space out the selection boxes and make their appearance more uniform and in line with each other. We increased the size of the confirm button and moved the alert messages to beneath the confirm bookiing button. We changed some of the labels and button text to more accurately reflect what they represented.

Resolves: #25